genhomedircon: add support for RBAC and %group expansion
This commit replaces the genhomedircon implementation with a slightly different
version that works with SELinux context_t objects directly, instead of looking
for instances of "system_u" in file context specifications it extracts the
context and replaces both the MLS level and user (and also role if RBAC is
enabled) with the respective values for the seuser mapping.

It also implements support for expansion of group names from seusers by finding
all members with a primary or supplementary group matching the given group name.
If a user is a member of 2 groups which are listed in seusers, this will be
treated as an error.

The support for RBAC works by replacing all roles in file context specifications
with whatever the user prefix is for the given SELinux user. The user prefix can
be set from policy in either a gen_user call, or a userprefix CIL statement.

refpolicy:

	gen_user(user_u, user_r, user_r, s0, s0)
                           ^
                   This is the users prefix
CIL:

	(user user_u)
	(role user_r)
	(userrole user_u user_r)
	(userprefix user_u user_r)

This feature can be toggled by setting "enable-genhomedircon-rbac" to True in
/etc/selinux/semanage.conf.

NOTE: This is mostly rewritten and will require picking out small changes to
upstream to libsemanage.